Cleanup
- cleanup documentation
- modularize PyTest command for DRY-ness and add to __init__ for accessibility
- reference module commands in setup.py
