+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 All notable changes to this project will be documented in this file.
+
 0.1.0 (2016-07-21)
 ------------------
 Initial release
