+++ pyreleaseplugin/__init__.py
@@ -1,2 +1,3 @@
 from pyreleaseplugin.clean import CleanCommand
 from pyreleaseplugin.release import ReleaseCommand
+from pyreleaseplugin.test import PyTest
